requirements.txt
ultralytics
pillow
opencv-python
albumentations
